16: moved pde tp utils to common testutils
Task-Url: http://github.com/LorenzoBettini/edelta/issues/16
