+++ edelta.parent/edelta.library.compilation/META-INF/MANIFEST.MF
@@ -16,7 +16,8 @@ Require-Bundle: org.junit;bundle-version="4.7.0",
  org.eclipse.xtext.testing,
  org.eclipse.xtext.xbase.testing,
  org.eclipse.xtext.ui.testing,
- org.eclipse.xtext.xbase.ui.testing
+ org.eclipse.xtext.xbase.ui.testing,
+ edelta.testutils
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Import-Package: org.hamcrest.core,
  org.junit;version="4.5.0",
