+++ edelta.parent/edelta.ui.tests/META-INF/MANIFEST.MF
@@ -18,7 +18,8 @@ Require-Bundle: edelta.ui,
  org.eclipse.xtext.testing,
  org.eclipse.xtext.xbase.testing,
  org.eclipse.xtext.ui.testing,
- org.eclipse.xtext.xbase.ui.testing
+ org.eclipse.xtext.xbase.ui.testing,
+ edelta.testutils
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: edelta.ui.tests;x-internal=true
 Import-Package: org.hamcrest.core,
