+++ edelta.parent/edelta.ui.tests/src/edelta/ui/tests/EdeltaNewProjectWizardTest.xtend
@@ -2,16 +2,16 @@ package edelta.ui.tests
 
 import com.google.inject.Inject
 import com.google.inject.Provider
+import edelta.testutils.PDETargetPlatformUtils
 import edelta.ui.tests.utils.EdeltaTestableNewProjectWizard
-import edelta.ui.tests.utils.PDETargetPlatformUtils
 import edelta.ui.tests.utils.PluginProjectHelper
 import org.eclipse.jface.viewers.StructuredSelection
 import org.eclipse.jface.wizard.Wizard
 import org.eclipse.jface.wizard.WizardDialog
 import org.eclipse.ui.PlatformUI
-import org.eclipse.xtext.ui.testing.AbstractWorkbenchTest
 import org.eclipse.xtext.testing.InjectWith
 import org.eclipse.xtext.testing.XtextRunner
+import org.eclipse.xtext.ui.testing.AbstractWorkbenchTest
 import org.junit.BeforeClass
 import org.junit.Test
 import org.junit.runner.RunWith
